backend/requirements.txt
streamlit==1.33.0
numpy
requests
pandas==2.2.2
streamlit_option_menu
pickle
matplotlib.pyplot
seaborn
fastapi
uvicorn
sklearn